+++ code/test/sixsq/nuvla/server/resources/nuvlabox_0_lifecycle_test.clj
@@ -617,7 +617,11 @@
                                                 :swarm-endpoint      "https://swarm.example.com"
                                                 :minio-access-key    "access"
                                                 :minio-secret-key    "secret"
-                                                :minio-endpoint      "https://minio.example.com"}))
+                                                :minio-endpoint      "https://minio.example.com"
+                                                :kubernetes-client-key    "key"
+                                                :kubernetes-client-cert   "cert"
+                                                :kubernetes-client-ca     "ca"
+                                                :kubernetes-endpoint      "https://k8s.example.com"}))
                 (ltu/body->edn)
                 (ltu/is-status 200))
 
@@ -629,10 +633,10 @@
                                         :body (rc/form-encode {:filter (format "parent='%s'" isg-id)}))
                                (ltu/body->edn)
                                (ltu/is-status 200)
-                               (ltu/is-count 2)
+                               (ltu/is-count 3)
                                (ltu/entries))]
 
-              (is (= #{"swarm" "s3"} (set (map :subtype services)))))
+              (is (= #{"swarm" "s3" "kubernetes"} (set (map :subtype services)))))
 
             ;; third commissioning of the resource make sure no additional credentials created
             (-> session
@@ -659,10 +663,10 @@
                                                                                isg-id)}))
                                (ltu/body->edn)
                                (ltu/is-status 200)
-                               (ltu/is-count 2)
+                               (ltu/is-count 3)
                                (ltu/entries))]
 
-              (is (= #{"swarm" "s3"} (set (map :subtype services))))
+              (is (= #{"swarm" "s3" "kubernetes"} (set (map :subtype services))))
 
               (doseq [{:keys [subtype] :as service} services]
                 (let [creds (-> session-owner
@@ -680,6 +684,9 @@
 
                   (if (= "s3" subtype)
                     (is (= 1 (count creds))))               ;; only key/secret pair
+
+                  (if (= "kubernetes" subtype)
+                    (is (= 1 (count creds))))
 
                   ))))